requirements.txt
# Document extraction
pdfplumber>=0.11.0
pymupdf>=1.24.0
python-docx>=1.1.0
mammoth>=1.8.0

# AI enrichment + config
requests>=2.32.0
pyyaml>=6.0.2

# CLI
rich>=13.7.0

# Web UI
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
